Move dom related functions to dom-cache.js
